Rearrange tests and package
